+++ src/main/java/kr/webgori/lolien/discord/bot/dto/LolienGenerateTeamDto.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class LolienGenerateTeamDto {
   private List<LolienSummoner> summonersTeam1;
   private List<LolienSummoner> summonersTeam2;
-  private int mmrDifference;
-  private int team1Mmr;
-  private int team2Mmr;
+  private float mmrDifference;
+  private float team1Mmr;
+  private float team2Mmr;
 }
